Add 'hidden' property for hosts
An hidden host is not returned via get on role, environment on all
hosts. It is however still present in the database, and returned if
called directly using the specific endpoint with the instance id.

This feature enable some hosts to be skipped during deployments.

Added a patch endpoint to update the hidden and terminable settings
for a host